refactors get_representations into networking.py
