+++ client/ayon_openrv/api/pipeline.py
@@ -7,23 +7,15 @@ import pyblish
 
 import rv
 
-from ayon_api import get_representations
-
 from ayon_openrv import OPENRV_ROOT_DIR
 
 from ayon_core.host import HostBase, ILoadHost, IWorkfileHost, IPublishHost
 from ayon_core.pipeline import (
-    load_container,
-    discover_loader_plugins,
-    get_current_project_name,
     register_loader_plugin_path,
     register_inventory_action_path,
     register_creator_plugin_path,
     AVALON_CONTAINER_ID,
 )
-
-from ayon_openrv import OPENRV_ROOT_DIR
-
 
 PLUGINS_DIR = os.path.join(OPENRV_ROOT_DIR, "plugins")
 PUBLISH_PATH = os.path.join(PLUGINS_DIR, "publish")
@@ -253,19 +245,3 @@ def get_containers():
         container = parse_container(node)
         if container:
             yield container
-
-
-def load_representations(representation_ids: list[str], loader_type: str):
-    """Load representations into current app session."""
-    project_name = get_current_project_name()
-
-    available_loaders = discover_loader_plugins(project_name)
-    if not loader_type:
-        raise ValueError("Loader type not provided. Expected 'FramesLoader' or 'MovLoader'.")
-    Loader = next(loader for loader in available_loaders
-                  if loader.__name__ == loader_type)
-
-    representations = get_representations(project_name, representation_ids=representation_ids)
-
-    for representation in representations:
-        load_container(Loader, representation, project_name=project_name)
