+++ client/ayon_openrv/networking.py
@@ -2,17 +2,19 @@ import json
 import socket
 from time import sleep, time
 
-from ayon_api import get_addon_settings
+from ayon_api import (
+    get_addon_settings,
+    get_representations
+)
+
+from ayon_core.pipeline import (
+    load_container,
+    discover_loader_plugins,
+    get_current_project_name,
+)
 
 from ayon_openrv.version import __version__
 from ayon_openrv.addon import OpenRVAddon
-
-try:
-    from ayon_openrv.api.pipeline import load_representations
-except ImportError:
-    # NOTE: import fails because we're running outside RV
-    #       if we launch RV from TrayLoader.
-    pass
 
 
 from ayon_core.lib.transcoding import (
@@ -216,3 +218,19 @@ class LoadContainerHandler:
             representation_ids = [i["representation"] for i in ayon_containers["MovLoader"]]
             log.debug(f"{representation_ids = }")
             load_representations(representation_ids, loader_type="MovLoader")
+
+
+def load_representations(representation_ids: list[str], loader_type: str):
+    """Load representations into current app session."""
+    project_name = get_current_project_name()
+
+    available_loaders = discover_loader_plugins(project_name)
+    if not loader_type:
+        raise ValueError("Loader type not provided. Expected 'FramesLoader' or 'MovLoader'.")
+    Loader = next(loader for loader in available_loaders
+                  if loader.__name__ == loader_type)
+
+    representations = get_representations(project_name, representation_ids=representation_ids)
+
+    for representation in representations:
+        load_container(Loader, representation, project_name=project_name)
